Move ConservedVars dataclass to fluid module.
